src/requirements.txt
requests
Flask
flask_session
pandas
cachetools
matplotlib
dash
dash-bootstrap-components
dash_bootstrap_templates
dash_cognito_auth
python-dotenv
humanize
dash-ag-grid
pydantic
pytz
unidecode
validators
markdown